Move exceptions to non-internal namespace
